fix: add @Transactional to our service methods and rename mapper functions for readability

+++ src/main/java/com/ibrawin/taskapi/services/TaskServiceImpl.java
@@ -8,6 +8,7 @@ import com.ibrawin.taskapi.model.TaskResponse;
 import com.ibrawin.taskapi.repositories.TaskRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -16,50 +17,56 @@ import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class TaskServiceImpl implements TaskService {
 
     private final TaskRepository taskRepository;
     private final TaskMapper taskMapper;
 
     @Override
+    @Transactional
     public TaskResponse saveTask(TaskRequest request) {
-        Task task = taskMapper.taskRequestToTaskMapper(request);
+        Task task = taskMapper.toTask(request);
         task.setCreatedAt(LocalDateTime.now());
         taskRepository.save(task);
-        return taskMapper.taskToTaskResponseMapper(task);
+        return taskMapper.toTaskResponse(task);
     }
 
     @Override
     public List<TaskResponse> getTasks() {
         return taskRepository.findAll()
                 .stream()
-                .map(taskMapper::taskToTaskResponseMapper)
+                .map(taskMapper::toTaskResponse)
                 .toList();
     }
 
     @Override
     public TaskResponse getTaskById(UUID id) {
         return taskRepository.findById(id)
-                .stream()
-                .map(taskMapper::taskToTaskResponseMapper)
-                .findFirst()
+                .map(taskMapper::toTaskResponse)
                 .orElseThrow(NotFoundException::new);
     }
 
     @Override
+    @Transactional
     public TaskResponse updateTaskById(UUID id, TaskRequest request) {
         Optional<Task> oldTask = taskRepository.findById(id);
         if (oldTask.isPresent()) {
-            Task newTask = taskMapper.taskRequestToTaskMapper(request);
+            Task newTask = taskMapper.toTask(request);
             newTask.setId(id);
             newTask.setCreatedAt(oldTask.get().getCreatedAt());
-            return taskMapper.taskToTaskResponseMapper(taskRepository.save(newTask));
+            return taskMapper.toTaskResponse(taskRepository.save(newTask));
         }
         throw new NotFoundException();
     }
 
     @Override
+    @Transactional
     public void deleteTaskById(UUID id) {
-        taskRepository.deleteById(id);
+        if (taskRepository.existsById(id)) {
+            taskRepository.deleteById(id);
+        }
+
+        throw new NotFoundException();
     }
 }
